requirements.txt
anaconda==0.0.1.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
beautifulsoup4==4.12.3
blinker==1.9.0
certifi==2025.7.9
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.4.2
click==8.2.1
cloud-init==25.1.4
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contourpy==1.3.2
cryptography==3.4.8
curl_cffi==0.11.4
cycler==0.12.1
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
exceptiongroup==1.3.1
fastapi==0.131.0
filelock==3.20.0
Flask==3.1.1
fonttools==4.58.5
frozendict==2.4.6
fsspec==2025.12.0
git-filter-repo==2.34.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.20.2
httpx==0.28.1
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
kiwisolver==1.4.8
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lxml==6.0.2
MarkupSafe==3.0.2
matplotlib==3.10.3
more-itertools==8.10.0
mpmath==1.3.0
multitasking==0.0.11
netifaces==0.11.0
networkx==3.4.2
numpy==2.2.5
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.6.77
oauthlib==3.2.0
openai==2.17.0
opencv-python==4.11.0.86
packaging==25.0
pandas==2.3.1
peewee==3.18.2
pillow==11.3.0
platformdirs==4.3.8
protobuf==6.31.1
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycparser==2.22
pycurl==7.44.1
pydantic==2.12.5
pydantic_core==2.41.5
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu4.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==5.4.1
requests==2.32.4
scikit-learn==1.7.2
scipy==1.15.3
SecretStorage==3.3.1
service-identity==18.1.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.52.1
sympy==1.14.0
systemd-python==234
threadpoolctl==3.6.0
tqdm==4.67.1
triton==3.5.1
Twisted==22.1.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
UNKNOWN @ file:///root/PyHazards
urllib3==1.26.5
uvicorn==0.41.0
wadllib==1.3.6
websockets==15.0.1
Werkzeug==3.1.3
yfinance==0.2.65
zipp==1.0.0
zope.interface==5.4.0
